Factor out references to UI modules
This is the first step toward moving to a background thread the code
that parses package files and extracts the key bindings.

The sublime and sublime_plugin modules are only available on the main UI
thread, so we need to collect relevant settings up front so they can be
passed into the background thread.