Show activity in status bar.
It takes a few seconds for the key bindings quick panel to appear, so
this code adds a little animated activity indicator in the status bar.

Since the status bar is only available if there's a view (i.e. tab), it
checks to make sure that a view exists before attempting to show the
animation.